b6e01988-4dd3-426e-8bc4-b5ed86f12a7e/requirements.txt
discord.py
dismusic
wavelink
discord-custom-help